atualizações 21/05

+++ Fabio-Aplication/Controllers/FuncionariosController.cs
@@ -1,5 +1,6 @@
 ﻿using Fabio_Aplication.Models;
 using System;
+using System.Data.Entity;
 using System.Linq;
 using System.Web.Mvc;
 
@@ -12,7 +13,10 @@ namespace Fabio_Aplication.Controllers
         // GET: Funcionarios
         public ActionResult Index()
         {
-            var funcionarios = db.FUNCIONARIOS.ToList();
+            var funcionarios = db.FUNCIONARIOS
+                           .Include(f => f.PAGAMENTO_FUNCIONARIOS)
+                           .ToList();
+
             return View(funcionarios);
         }
 
@@ -31,12 +35,10 @@ namespace Fabio_Aplication.Controllers
             {
                 funcionario.DATACAD = DateTime.Now;
                 funcionario.DATAALT = DateTime.Now;
-
                 db.FUNCIONARIOS.Add(funcionario);
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
-
             return View(funcionario);
         }
 
@@ -45,13 +47,11 @@ namespace Fabio_Aplication.Controllers
         {
             var funcionario = db.FUNCIONARIOS.Find(id);
             if (funcionario == null)
-            {
                 return HttpNotFound();
-            }
             return View(funcionario);
         }
 
-        // POST: Funcionarios/Edit/5
+       
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Edit(FUNCIONARIOS funcionario)
@@ -59,31 +59,34 @@ namespace Fabio_Aplication.Controllers
             if (ModelState.IsValid)
             {
                 funcionario.DATAALT = DateTime.Now;
-
-                db.Entry(funcionario).State = System.Data.Entity.EntityState.Modified;
+                db.Entry(funcionario).State = EntityState.Modified;
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
             return View(funcionario);
         }
 
-        // GET: Funcionarios/Delete/5
         public ActionResult Delete(int id)
         {
             var funcionario = db.FUNCIONARIOS.Find(id);
             if (funcionario == null)
-            {
                 return HttpNotFound();
-            }
             return View(funcionario);
         }
 
-        // POST: Funcionarios/Delete/5
+      
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
             var funcionario = db.FUNCIONARIOS.Find(id);
+            if (funcionario == null)
+                return HttpNotFound();
+
+            
+            var pagamentos = db.PAGAMENTO_FUNCIONARIOS.Where(p => p.FUNCIONARIO_ID == id).ToList();
+            db.PAGAMENTO_FUNCIONARIOS.RemoveRange(pagamentos);
+
             db.FUNCIONARIOS.Remove(funcionario);
             db.SaveChanges();
             return RedirectToAction("Index");
